Feat: Add EnumTypeEntity.java
