--- api/src/main/java/com/thesurvey/api/domain/CertificationType.java
@@ -1,9 +1,0 @@
-package com.thesurvey.api.domain;
-
-public enum CertificationType {
-    KAKAO,
-    NAVER,
-    GOOGLE,
-    WEBMAIL,
-    DRIVER_LICENSE
-}
