+++ api/src/main/java/com/thesurvey/api/domain/EnumTypeEntity.java
@@ -1,0 +1,25 @@
+package com.thesurvey.api.domain;
+
+public class EnumTypeEntity {
+    public enum QuestionType {
+        SINGLE_CHOICE,
+        MULTIPLE_CHOICE,
+        SHORT_ANSWER,
+        LONG_ANSWER
+    }
+
+    public enum CertificationType {
+        KAKAO,
+        NAVER,
+        GOOGLE,
+        WEBMAIL,
+        DRIVER_LICENSE
+    }
+    public enum Role {
+        USER,
+        ADMIN
+    }
+
+
+
+}
