--- api/src/main/java/com/thesurvey/api/domain/QuestionType.java
@@ -1,8 +1,0 @@
-package com.thesurvey.api.domain;
-
-public enum QuestionType {
-    SINGLE_CHOICE,
-    MULTIPLE_CHOICE,
-    SHORT_ANSWER,
-    LONG_ANSWER
-}
