--- api/src/main/java/com/thesurvey/api/domain/Role.java
@@ -1,6 +1,0 @@
-package com.thesurvey.api.domain;
-
-public enum Role {
-    USER,
-    ADMIN
-}
